Working
Note Synchronization works.

Must create a midi file set up with channels 1-6 acting as each button.

Must pad the start of the audio file by 4 beats with white space, since we use audio_sound_get_track_position() to spawn the notes, the white space is to ensure you can see the notes falling before the song starts to prepare

in non_gamemaker_files/ there is a midi template for use in Renoise, the Renoise file for the first song, and a python script to convert the raw MIDI data to JSON