Added test mode
game will activate the "test_mode_auto_play" object if a global flag is set in the song select menu by holding down left and right at the same time. visual added to bottom left to show when the flag is enabled or not.

when activated, the game will essentially play itself.

also adjusted the note speed calculation slightly